Seems to work
One of problems is how difficult it is to write tests for controller
Parsing and processing of messages are in one place and not shown in
interface. I wasted a lot of time during weekend thinking how to fix
that. In fact, barely had time to finish a working version.

If I have time, I'll push a refactored version on another branch